WpfPlot as CustomControl
converted the UserControl to a CustomControl to facilitate inheritance and theming
