+++ src/ScottPlot5/ScottPlot5 Controls/ScottPlot.WPF/Themes/Generic.xaml
@@ -3,13 +3,22 @@
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:local="clr-namespace:ScottPlot.WPF"
     xmlns:skiaPreview="clr-namespace:SkiaSharp.Views.WPF"
+    xmlns:skia="clr-namespace:SkiaSharp.Views.WPF;assembly=SkiaSharp.Views.WPF"
     >
-
     <Style TargetType="{x:Type local:WpfPlotGL}">
         <Setter Property="Template">
             <Setter.Value>
                 <ControlTemplate TargetType="{x:Type local:WpfPlotGL}">
                     <skiaPreview:SKGLElement x:Name="PART_SKElement" />
+                </ControlTemplate>
+            </Setter.Value>
+        </Setter>
+    </Style>
+    <Style TargetType="{x:Type local:WpfPlot}">
+        <Setter Property="Template">
+            <Setter.Value>
+                <ControlTemplate TargetType="{x:Type local:WpfPlot}">
+                    <skia:SKElement x:Name="PART_SKElement" />
                 </ControlTemplate>
             </Setter.Value>
         </Setter>
